Remove some global state; Separate game init from run
Do not add the dungeon, map, or rooms to the global game object
Screen, Keyboard, and Assets are still global object literals
to guarantee that there are never multiple instances
Player also has to be a global because somehow passing parameters to the
main game loop blows the call stack
The game's init function now only generates game objects like the
dungeon and map. A new run function loads the actual images, draws
to the screen, and starts the game.