Create new tiles, draw walls, random dungeon colors
The new tiles all have a grayscale version and a black version for
contrast. By splitting tiles and layering them, I can add a
semi-transparent fill for random dungeon colors. Without the contrast of
the black, the colors were too washed out.
The tiles include walls which are now being drawn, and doors (both open
and locked) which are not yet being drawn.
The tiles also include a water tile, which was tested with a new room
layout, 'brackets'.