First pass at drawing doors
Currently reading the room object and drawing doors in the correct
locations. The transparency is causing the doors to have lines
through them though. In any case they need to be split into sub-layers
and drawn on another canvas so that Link can appear to move underneath
the top portion of the wall.