Link should appear to walk under door frames
Move player to new main canvas.
Use fg canvas for frame around walls and for section above doors.
This canvas is above the main canvas so that Link appears to walk
underneath.

Also, move the random color generator to the dungeon file so that it
uses the dungeon seed value.